mutt: Change email checking and notification
Do not send notifications for new emails. This will be done by another
program at some point.

Also, only check for new mail every minute if open. Most of the time
mutt is not supposed to be open but in case it is, it won't be
requesting attention so often.